# openapi/api.yaml
openapi: "3.0.3"
info:
  title: DragonHorde
  version: "1.0"
servers:
  - url: https://api.server.test/v1
paths:
...
  /media/{id}:
    get:
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of requested media
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "schema/media.json"
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      tags:
        - media               
    put:
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of requested media
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "schema/media.json"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "schema/media.json"
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      tags:
        - media                
    delete:
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of requested media
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      tags:
        - media
...